14.RBAC模型附加功能，15.权限拦截块

+++ p1xn6d2/permission/src/main/java/com/mmall/service/impl/SysAclModuleServiceImpl.java
@@ -2,6 +2,7 @@ package com.mmall.service.impl;
 
 import com.google.common.base.Preconditions;
 import com.mmall.common.RequestHolder;
+import com.mmall.dao.SysAclMapper;
 import com.mmall.dao.SysAclModuleMapper;
 import com.mmall.exception.ParamException;
 import com.mmall.model.SysAclModule;
@@ -27,6 +28,8 @@ import java.util.List;
 public class SysAclModuleServiceImpl implements SysAclModuleService {
 
   @Resource private SysAclModuleMapper sysAclModuleMapper;
+
+  @Resource private SysAclMapper sysAclMapper;
 
   @Override
   public void save(AclModuleParam param) {
@@ -120,9 +123,9 @@ public class SysAclModuleServiceImpl implements SysAclModuleService {
     if (sysAclModuleMapper.countByParentId(aclModule.getId()) > 0) {
       throw new ParamException("当前模块下面有子模块，无法删除");
     }
-    /*     if (sysAclMapper.countByAclModuleId(aclModule.getId()) > 0) {
-        throw new ParamException("当前模块下面有用户，无法删除");
-    }*/
+    if (sysAclMapper.countByAclModuleId(aclModule.getId()) > 0) {
+       throw new ParamException("当前模块下面有用户，无法删除");
+    }
     sysAclModuleMapper.deleteByPrimaryKey(aclModuleId);
   }
